refactor[pages]: page hierarchy
