+++ client/src/pages/user/UserBlogsPage/index.jsx
@@ -1,14 +1,14 @@
 import { Modal, Select } from 'antd';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useParams, useSearchParams } from 'react-router-dom';
-import { deleteBlog } from '../../api/services/blogService';
-import BlogList from '../../components/BlogList';
-import Error from '../../components/Error';
-import useApiErrorHandler from '../../hooks/useApiErrorHandler';
-import useAuth from '../../hooks/useAuth';
-import useConfirm from '../../hooks/useConfirm';
-import useFeedback from '../../hooks/useFeedback';
-import useFetch from '../../hooks/useFetch';
+import { deleteBlog } from '../../../api/services/blogService';
+import BlogList from '../../../components/BlogList';
+import Error from '../../../components/Error';
+import useApiErrorHandler from '../../../hooks/useApiErrorHandler';
+import useAuth from '../../../hooks/useAuth';
+import useConfirm from '../../../hooks/useConfirm';
+import useFeedback from '../../../hooks/useFeedback';
+import useFetch from '../../../hooks/useFetch';
 import styles from './UserBlogsPage.module.css';
 
 const selectOptions = [
